change hockey stick to a timeseries chart.
requires underlying data to be represented as dates, so moved from '1000', etc. to '1000-01-01', etc.

suggest that this might be an option for expanding the easy-to-use graph types... making a series of parameterised includes for a range of 'styles', rather than making one _CRAZY_ API; or exposing the authors to the almost-infinite configurability of c3/d3